Fixed issue with Profile
Made the profile code use fragments so that we can re-use it for the
authenticated user's profile and for any friend's profile.
1) Refactored ProfileFragment into an abstract base class.  Implement
AuthenticatedProfileFragment and UserProfileFragment using it.
1) Renamed ProfileActivity to AuthenticateProfileActivity, since it is
the profile for the authenticated user.  Updated activity_profile.xml to
use the authenticated user fragments.
2) Created UserProfileActivity which uses UserProfileFragment and
UserTimelineFragment via activity_user_profile.xml.
3) Updated TweetsAdapter to listen for clicks on the image and start the
profile activity.